Make the arm kernel run on the Raspberry Pi 3.
It seems that the preferred rpi3 kernel for mainline source
is 64-bit.  So to get a 32-bit kernel to work I followed
the suggestions at this page:

https://wiki.gentoo.org/wiki/Raspberry_Pi/Mainline_Kernel

which is essentially this: configure with multi_v7_defconfig,
use the device tree from arm64.

It turns out that the configuration is very similar to what
we were using already, so the only change is to configure
the 8250 serial port correctly so we get serial console.

The device tree change is done with a patch that I hand crafted.

Also, fix small typo in Makefile so we actually compile
the device tree files.