Get rpi wlan to work (but still must be started manually).
The biggest change is to use a newer kernel (4.13.4) that has
better support for the hardware, including new device trees
and firmware.  For one thing, it no longer needs a patch.

Add kmod and wpasupplicant packages to the build.

Remove rootfs.fakeroot when doing rootfs_clean.  It's not
necessary, just nice.