Turn on wlan automatically.
For now, the interface must be wlan0.

If a file /boot/wlan.conf exists, then it is used to generate
wpa_supplicant.conf, then wpa_supplicant and dhclient are started
on wlan0.

wlan.conf consists of lines of the form "ssid psk".  Quoting
with " is allowed and recommended.  Use \" to embed a literal
quote character (but only in a qouted string).  (Yes, people
can and some of them do use whitespace in their SSIDs.)
Comments beginning with # are allowed in the usual way.