Add optional devel directory for extra files to install.
If devel/rootfs exists, its contents are copied into rootfs
with hierarchy preserved.

If devel/imgfs exists, its contents are copied into the boot
partition (also with hierarchy preserved).  This is a good way
to preload the image with a wlan configuration.

I also renamed build/.../img_fs to build/.../imgfs, to be
consistent with rootfs (and because _ is hard to type).

Change the building of (pc and rpi) images to copy files recursively.

Stop $(PC_IMG) from depending on mkfatimg.  It doesn't use it.