Add ntpd.
Or more specifically, openntpd, which is a much smaller package.

Currently, we are using debian.pool.ntp.org, which is the default.

In startwlan, start dhclient without the -nw flag so the network
is ready to use before the script exits.  However, if there is
no suitable wireless network available, then it can block for
a long time (followed by ntpd blocking for a long time for
a response so it can set the initial time).