Add docker support.
This required adding a bunch of new kernel config options. While
doing that, I added some security hardening options and cleaned
up the kconfigs a bit.

The new options made the uncompressed kernel too big to fit, so
it clobbered the dtb and initramfs.  So I changed the uboot env
to make more room (up to 32MB).