Add default hostname and /etc/hosts.
We use an initial default hostname, then set it once the first
time we get one from dhcp (if we do).  Because the normal
dhclient mechanism won't override the existing hostname unless it's
something like "(none)", we have to use our own script to do it.

Docker doesn't export /etc/hosts or /etc/hostname so we have
to put them in ourselves in make.

The hosts file is the standard one from debian or ubuntu or
one very similar to it.