Switch root from initramfs to a squashfs during boot.
/init is now split into two parts.  The first part (/init)
performs early initialization, mounts the necessary filesystems,
and execs "switch_root ... /init1".  The second part (/init1)
is essentially the tail end of the old /init.

For now, both roots have the same contents.

/boot is now mounted read-only because it can never be
unmounted (while root is mounted from a file under it)
and we don't want it to be marked unclean on reboot.
When necessary, it can be temporarily remounted read-write
to make changes.

Because the new root is read-only (as a squashfs always is),
/run and /tmp are now tmpfs mounts and parts of / are links
into /run so then can be changed.

Build-time customization of rootfs after it's exported from
docker is now in a single script, instead of split up in
makefile commands and the mkdevnodes script.