Fix amd64 image.
The amd64 kernel didn't compile in squashfs or HFS.

Also, fstab had a hard-coded path to the sdcard, I changed this to
use LABEL= mount, which required setting a label on the rpi3 FAT fs.
That label is "ISOIMAGE" to match what grub-mkrescue generates on
the PC side.  In theory, we could change the label on the HFS+
filesystem to something that makes more sense, but there are no
off the shelf tools for that, and HFS+ is pretty complicated.