Cleanup docker image build.
Switch to using build-args instead of faking Dockerfile templating
with sed.

Use tar to collect artifacts (and the Dockerfile) so we can pull
files both from the rootsrc dir and from the build/ subtree.

Finally, move the list of extra .debs to download into a dedicated
file, not the Makefile.