Clean up building of busybox initrd.
Keep initrd files in a directory and use a dockerfile.
I think this is more in keeping with the spirit of things.

Move the real root's init into /sbin.  It was called /init1;
now it's /sbin/init.

Use a fixed version of the busybox container (1.28.0) instead of
the latest.  Also, explicitly use the uclibc version instead of
just defaulting to it.

Clean up makefile rules for removing docker residues.