Add rudimentary storage management.
We use busybox mdev, which invokes the new script, storagemgr,
when a block device partition is added.  For now, if the partition
is mountable and contains files of the form pv[0-9][0-9][0-9]
in the directory /prjx/storage, then they are considered
LVM physical volumes.  The physical volumes together should
make up a volume group named vg0, which should contain a logical
volume named lv0.  When all the pieces have been assembled,
vg0/lv0 (if it contains a filesystem) will be mounted at
/run/mnt/vg0-lv0.

Unplugging a drive doesn't really work yet.

The PC version builds and runs, but the storage part
is untested.  Also, the PC boot drive is sda, and storagemgr
needs to know not to touch it.  It's OK for now because
it's an ISO image so doesn't have any partitions and
storagemgr only acts on partitions.

Associated changes:

Add device mapper to kernel configuration.

Add lvm to the rootfs docker.

Add minimal mdev configuration to invoke storagemgr and
to help LVM along.

There is no such thing as a simple shell script.