Generalize how we build containers.
This changes how we build containers to both make rootfs less special,
and to allow multiple containers to be build and loaded into docker.

The idea is that for each container we build, we make a squashfs
for each layer.  We also save the tar fragments needed to recreate
the image tar for loading into docker.  Finally, we save the list
of layers for each container, in order.  We mainly only need this
for the rootfs, since the initrd needs to overlayfs the real root,
so it needs to know the names.

This requires enabled "experimental" mode on your docker deamon.  The
makefile will warn if this is not done.