Add interim storage provisioning function to storagemgr.
This is not exactly on the path to the real stuff, but it's
a convenient feature for now.

Run "storagemgr -p dev" (where dev is a partition, like sda1)
to create a single 4G file on that device, make it into an lvm
physical volume that's part of a volume group with a single
logical volume, make an ext4 filesystem on that logical volume,
and mount it.

Collateral damage:

Mount points are now in /mnt instead of /run/mnt, because
eventually it'll have to be mounted shared with a container
and it's nice to limit the damage.

All the lvm names now have an additional numeric digit: pv0000,
vg00, lv00.