Fix brcm firmware download and installation.
We get the Broadcom firmware from the rpi github repository,
and they seem to have changed things on us.  This makes it
work again.

The download size and the extracted firmware directory size are
both bigger, which is not desirable.  We should eventually get
the firmware via apt-get in the normal way.