Switch to storagemgr in python.
The new code is roughly functionally equivalent to the old.

Creating the initial storage (called provisioning, provisionally)
remains a problem.  It takes a long time, runs in a container,
and, if done automatically, would be called from a daemon.
I don't know how to coordinate this asychronous operation
with the synchronous parts of init.
So it's not happening, yet.

I negated the meaning -d in both versions of storagemgr.
It now means daemon (log to syslog) instead of debug
(don't log to syslog).