Patch out fallocate support in fatfs.
ext4 resizing uses fallocate options that are not supported
by fatfs.  It has a fallback strategy but only if fallocate
is not supported at all, not if it fails after being called.
So for now, we just take it out altogether.

Also, loop.c has to pass EOPNOTSUPP up the layers, instead of
squashing it into EIO.