Add lvm file growing to storagemgr
It's still not done automatically.  That will be coordinated
by a shell script on the host.  The manual procedure:
   smdocker -psda1 -s128 -v
   smdocker -gsda1 -v
The first command provisions the drive (with one 128MB file),
the second grows the file and creates new ones until the drive
is mostly full.  Add -f<n> to limit the growth to <n> files,
-s<size> to set the file size (default 4000MB), and -m<minfree>
to change the reserved fs free space (default 500MB).
To run in the background, do "smdocker -d ..." and it'll
start the container in detached mode and storagemgr will
log to syslog.

And a few small changes like using "docker run -t" for running
storagemgr in the foreground.