Rudimentary DNS server.
dnsd is a simple DNS server based on python dnslib.
It runs in a container connected to both the external network
and the zerotier private network and serves both an external
zone and an internal one.

For now, it has only a dummy zone database.  The real zones
need to be populated by a script that knows the services
available and the external address of the host.