Add control channel to dnsd, restructure code.
There is for now exactly one control command: set value.
It is also not persistent across reboot, so it's only suitable
for temporary values, like the TXT token for ACME dns-01
authorization.

The protocol is a single json exchange: request as a vector
of command and arguments, and reply as a vector of status code
and string message.  We can make it something a little more
elaborate if we want.

The code is substantially restructured, cleaned up, and made
more robust.  For example, instead of dying horribly on
an exception, server threads now catch the exception, log it,
and return SERVFAIL.  Also, the zone data is now locked
because it can be modified by the main thread while being
used by the dns server threads.