dnsd rework.
Reworked dnsd to be more like what we want.

Only one zone is supported.  It is initially empty, and should
be populated via the control channel.  The commands are:

["init", origin, external-address, ttl]
    origin: (string) the base domain name
    extern-address: (string) the external address of the host
    ttl: (int) TTL for all records

    Initialize the zone with SOA, NS, and A records on origin,
    and an A record for ns.origin.

["add", name, type, data]
    name: (string) the name to be added
    type: (string) "a", "mx", "cname", etc.
    data: (depending on type) the resource data

    Add a record for name, which can be absolute (ending in "."),
    or relative to origin (including "@" for the origin itself).
    If data is an address, then "@" means the address of origin
    (the external address).

["delete", name, type]
    name: (string) the name to be added
    type: (string) "a", "mx", "cname", etc.

    Delete all records matching name and type.  Type may be "any".

The database is not presistent.

Name and type are case insensitive.  Name is always stored
in lowercase, but replies try to preserve the case in the request
(per rfc1034, rfc1035).

The zone can be initialized and populated with data in a file
specified on the command line, but I expect this to be used only
for debugging.

The dns server listens on the wildcard address.  The control channel
is only reachable on the private address.  I'll probably have to
change how the latter is determined.