Sort out dnsd addresses one last time.
A private network "dnsdcontrol" is created for the control
socket.  The dnsd container is attached to it with
"docker network connect --alias=dnsd-control".
Internally, dnsd listens on the name "dnsd-control" for commands.
Any other container attached to the same network can talk to it
using the same name.

Bonus change: Wrap the init command in a catch frame so it can
clean up a failed initialization.