Certificate manager (certmgr)
OK.  This thing runs in a container, executes certbot, and
talks to dnsd via the control channel to implement dns-based
authentication.

Each certmgr execution handles exactly one request-reply
exchange.  It takes requests from stdin (in json) or
on the command line.  The two supported commands are
new and renew.

On the command line,

    certmgr new domain
    certmgr renew domain

In json,

    ["new", <domain>]
    ["renew", <domain>]

The reply is always in json:

    [0, [<cert>, <key>]]

<cert> is the certificate (full-chain) and <key> is the private key.
They are strings.

Error reply is

    [<error>, <msg>]

<error> is a nonzero integer and <msg> is a string.

The only persistent state for the container is the registered
letsencrypt user account, which is created on demand and reused.
When certmgr runs, a certificate is created, passed back to the
caller, then goes poof when the container exits.

Because we don't keep track of certificates, new and renew
actually do exactly the same thing.  I believe letsencrypt
doesn't differentiate them either.

There's more to do, like the interface with app store and
the framework for renewing soon-to-expire certificates.
The work to invoke the container is a little involved
(connect dnsdcontrol network, mount persistent storage).
It's currently contained in scripts on the host, but
in production certmgr will be started by the app store
container.  We should have a way to encapsulate that.