Add 'appstore' container.
Note that this has a ridiculous hack for importing the containers
into docker.  The sane approach of streaming from the HTTP download
to the docker daemon's socket times out frequently due to the docker
daemon not taking bytes in time.  So for now we download to the
USB storage, and then a daemon (written in bash) on the host looks
for them periodically and "docker load"s them.

There is a related hack for starting the app containers, which means
apps won't automatically restart on later boots.