Put private zerotier network on the built-in controller.
This will ultimately make it easier to bring up the private
network, since everything is now under our control.

For now, the way to add hosts to the private network is
to put their zerotier host ids in zerotier.conf like this:
    ZEROTIER_AUTHORIZED_HOSTS="<host-id>..."
Then run "zerotier join <network-id>" on each host after
the device boots and prints the network id on the console.