Client based pairing.
The pi now registers w/ boundery.me central, and joins a temporary
public ZT network.

The client also registers w/ boundery.me, then pairs w/ the pi
via the temporary public network.  This is secured with a shared
secret.

The pi and the client then both join a secure private ZT network.

The client currently hard-codes the shared secret to 32 'X's, which
goes into a file in /boot called "pairingkey".  Later, the client
will be extended to support flashing the SD itself, so it can write
the shared key then.