Port to 64-bit ARM
The armhf build is still available as an option, for now.

u-boot config is just rpi_3_defconfig.  Version was bumped
up to 2017.11.

Kernel config is a pared-down version of defconfig.
Version is unchanged.

Compressed kernel is not supported an arm64, so for now it's
uncompressed.  If we want to, the kernel can be packaged with
initrd and dtb into a u-boot image, which can be compressed.

For some reason logger doesn't support --id anymore, so I had
to stop using it.

zerotier-one has to be compiled from scratch and linked
dynamically.  The statically linked version crashes
mysteriously in pthreads (on both arm64 and amd64), when
a network is created and joined.  I had known this on amd64
for a while now, but it wasn't a problem on armhf.
Instead building a zerotier debian package and installing
it in the build container, I chose to do the cheesy thing
of installing the standard package and copying our compiled
zerotier-one binary over it.  I feel only a little ashamed.

The latest version of docker forced some unrelated changes:
1. It's bigger so the boot image size has to be bigger.
2. "docker info" output format is different.
This happened today while I was preparing this commit,
so they get thrown in.