Fix amd64 booting.
Adjust partition labels so initrd and root see the same ones for
/boot.

Plant a spacer file in the imgfs at build time that can be deleted at
runtime to make space for writing apikey.

Use "force" when remounting /boot read-write, because the hfs+ FS that
grub-mkrescue creates has a journal, which Linux won't mount
read-write otherwise.