test run a cross compiled program instead of checking binfmt_misc
when building in a docker container, for example, binfmt_misc isn't already mounted by default, so the binfmt_misc check
in the Makefile fails. we actually rather be testing whether the host can run cross-compiled programs, which is what we
really want to know, especially because we don't care whether they're using qemu or something else to run them.