arm64: fix sdcard on rpi3 on linux >= 5.6
upstream commit 738987a1d6f1, "mmc: bcm2835: Use dma_request_chan() instead
dma_request_slave_channel()" causes sdhost-bcm2835 (which is used for the sdcard
on rpi3) to defer probing until DMA can be used. however, we had DMA_BCM2835
configured as a module on arm64.

the fix is either to change DMA_BCM2836 to built-in, or include the module in
the initrd. our armhf kconfig uses it as a built-in, so let's just do the same
for arm64.